doniServer/serializers/blogs/tag.py
from rest_framework import serializers
from doniServer.models import Tag


class TagSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Tag
        fields = ('name', 'url')