requirements.txt
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.1
decorator==4.0.10
Django==1.10.2
django-cities==0.5
django-filter==0.15.3
django-model-utils==2.6
django-rest-auth==0.8.2
djangorestframework==3.5.1
djangorestframework-jwt==1.8.0
enum34==1.1.6
get==0.0.0
ipython==5.1.0
ipython-genutils==0.1.0
Markdown==2.6.7
MySQL-python==1.2.5
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
post==0.0.0
prompt-toolkit==1.0.8
ptyprocess==0.5.1
public==0.0.0
Pygments==2.1.3
PyJWT==1.4.2
query-string==0.0.0
requests==2.11.1
selenium==3.0.1
setupfiles==0.0.13
simplegeneric==0.8.1
six==1.10.0
swapper==1.0.0
torctl==0.2
tqdm==4.10.0
traitlets==4.3.1
wcwidth==0.1.7
